analytics/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
ecdsa==0.19.0
fastapi==0.115.2
h11==0.14.0
idna==3.10
mailgun==0.1.1
pillow==11.0.0
pyactiveresource==1.0.1
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
python-dotenv==1.0.1
python-jose==3.3.0
reportlab==4.2.5
requests==2.32.3
rsa==4.9
schedule==1.2.2
six==1.16.0
sniffio==1.3.1
starlette==0.40.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
